requirements.txt
paho-mqtt
requests
python-slugify
urllib3

